Fix Ruby 3 compatibility
The bulk of the changes are stolen from a previous pull request authored
by Sammy requests that was never merged.

This does a few things that may not be necessary to fix compatibility:

In addition to the changes that Sammy originally proposed I also opted
to (1) Remove the Gemfile.lock file, and (2) bump a few dependencies,
due to some various issues connected to the ruby version.