Support instantiating different API clients
This allows you to use Tarpon to target multiple different RevenueCat
projects from within the same codebase.

This changes the library so that it now instantiates `Tarpon::Client`
and calls methods on the instances, rather than calling singleton
methods on the class.

In order to preserve backwards compatibility—and to support the common
use case of just needing to target a single RevenueCat project, this
also introduces a `Tarpon::Client.default` instance to which all methods
are delegated.

This allows users to continue using the same API, but provides the
flexibility to target different projects if needed.